moved .h files back into src

+++ Engine/src/Tools/StopWatch.h
@@ -3,15 +3,14 @@
 
 class StopWatch {
 private:
-	size_t m_laps;
+	int m_laps;
 	std::chrono::high_resolution_clock::time_point m_startTime;
 public:
 	StopWatch();
 
 	void start();
 	void stop();
+
 	std::chrono::duration<double, std::milli> getDuration();
 	std::chrono::duration<double, std::milli> getAverageLapDuration();
-
-
 };
